requirements.txt
Django==4.2.9
python-dotenv==1.0.1
python-telegram-bot==13.15